+++ frontend/src/components/molecules/ReportCard/BackReportCardHeader.tsx
@@ -1,6 +1,6 @@
 import CircleImage from '../../atoms/Image/CircleImage.tsx';
-import MinutesAgo from '../MinutesAgo_delete/MinutesAgo.tsx';
 import { css } from '@emotion/react';
+import CommonText from '../../atoms/Text/CommonText.tsx';
 
 interface propsState {
     minutesAgo: number;
@@ -10,10 +10,18 @@ export default function BackReportCardHeader({ minutesAgo }: propsState) {
     return (
         <div css={backReportCardHeaderStyle}>
             <CircleImage size='3.75rem' />
-            <MinutesAgo value={minutesAgo} />
+            <CommonText {...minutesAgoProps}>{`${minutesAgo}분전`}</CommonText>
         </div>
     );
 }
+
+const minutesAgoProps = {
+    TextTag: 'span',
+    fontSize: 'caption',
+    fontWeight: 'medium',
+    color: 'greyOpacityWhite-60',
+    lineHeight: 1.3,
+} as const;
 
 const backReportCardHeaderStyle = css`
     width: 23rem;
